refactor(app): unify key package upload into a single task
Merge the plain and APQ key package upload timed tasks into one task in
outbound_service/key_packages.rs. Ref bookkeeping moves to
key_stores/key_package_refs.rs. Orphaned key packages are removed only
after both the plain and the APQ half of a batch are uploaded, closing
the window where one upload could delete the other kind's bundles.
