+++ coreclient/src/key_stores/mod.rs
@@ -41,6 +41,7 @@ use serde::{Deserialize, Serialize};
 
 pub(crate) mod as_credentials;
 pub(crate) mod indexed_keys;
+pub(crate) mod key_package_refs;
 pub(crate) mod queue_ratchets;
 
 // For now we persist the key store along with the user. Any key material that gets rotated in the future needs to be persisted separately.
